Update the matrix of Ruby/Rails versions to test
The rule of thumb to test a gem against different combinations of Ruby and Rails should be: That the gem should pass tests against the oldest recommended Ruby version for a supported Rails version. Furthermore it should pass for all newer versions that are supported by that Rails version.

Therefore the matrix shoould be:

* Rails 3.2 - Ruby 1.9.3 (recommended)
* Rails 3.2 - Ruby 2.1.6
* Rails 3.2 - Ruby 2.2.2 (supported since 3.2.22)
* Rails 4.1 - Ruby 2.1.6 (recommended)
* Rails 4.1 - Ruby 2.2.2
* Rails 4.2 - Ruby 2.1.6
* Rails 4.2 - Ruby 2.2.2 (recommended)